Move makeMergedSchema
