+++ packages/graphql-server/src/__tests__/makeMergedSchema.test.ts
@@ -5,13 +5,13 @@ import {
   makeDirectivesForPlugin,
   createTransformerDirective,
   createValidatorDirective,
-} from '../../directives/makeDirectives'
+} from '../directives/makeDirectives'
+import { makeMergedSchema } from '../makeMergedSchema'
 import {
   GraphQLTypeWithFields,
   ServicesGlobImports,
   SdlGlobImports,
-} from '../../types'
-import { makeMergedSchema } from '../makeMergedSchema'
+} from '../types'
 
 jest.mock('@redwoodjs/project-config', () => {
   return {
