+++ packages/graphql-server/src/functions/graphql.ts
@@ -14,7 +14,7 @@ import { getConfig } from '@redwoodjs/project-config'
 import { mapRwCorsOptionsToYoga } from '../cors'
 import { makeDirectivesForPlugin } from '../directives/makeDirectives'
 import { getAsyncStoreInstance } from '../globalContext'
-import { makeMergedSchema } from '../makeMergedSchema/makeMergedSchema'
+import { makeMergedSchema } from '../makeMergedSchema'
 import {
   useArmor,
   useRedwoodAuthContext,
