+++ packages/graphql-server/src/index.ts
@@ -6,13 +6,14 @@ export * from './globalContext'
 export * from './errors'
 export * from './functions/graphql'
 export * from './functions/useRequireAuth'
-export * from './makeMergedSchema/makeMergedSchema'
+export * from './makeMergedSchema'
 export * from './types'
 
 export {
   createValidatorDirective,
   createTransformerDirective,
   getDirectiveName,
+  makeDirectivesForPlugin,
 } from './directives/makeDirectives'
 
 export {
@@ -26,6 +27,7 @@ export {
   TransformerDirectiveFunc,
   ValidateArgs,
   TransformArgs,
+  useRedwoodDirective,
 } from './plugins/useRedwoodDirective'
 
 export * as rootSchema from './rootSchema'
